+++ testData/org/elixir_lang/code_insight/parameter_info/beam/elixir_default_arguments.ex
@@ -1,0 +1,5 @@
+defmodule ParameterInfo.Beam do
+  def run do
+    Code.eval_string(<caret>)
+  end
+end