+++ testData/org/elixir_lang/code_insight/parameter_info/beam/erlang_pattern_parameters.ex
@@ -1,0 +1,5 @@
+defmodule ParameterInfo.Beam do
+  def run do
+    :queue.in(<caret>)
+  end
+end