+++ testData/org/elixir_lang/code_insight/parameter_info/beam/source_over_beam_declaration.ex
@@ -1,0 +1,3 @@
+defmodule Code do
+  def fetch_docs(module_or_path), do: module_or_path
+end